ecommerce-api/requirements.txt

# =====================================================
# Dépendances pour API E-commerce FastAPI
# Python 3.11+
# =====================================================

# FastAPI & ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.25
aiomysql==0.2.0
asyncmy==0.2.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis (Cache & Sessions)
redis==5.0.1
aioredis==2.0.1

# HTTP Client (pour webhooks, API externes)
httpx==0.26.0
aiohttp==3.9.1

# Image Processing (optionnel pour upload images)
Pillow==10.2.0

# Background Tasks
celery==5.3.4
flower==2.0.1

# Payment Integrations (exemples)
stripe==7.12.0
# paypalrestsdk==1.13.1  # Si besoin PayPal

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# Monitoring & Logs
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.39.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.4.0

# Development Tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Utilities
python-slugify==8.0.1
pytz==2023.3
cryptography==42.0.0